requirements.txt
### dependency
streamlit
nltk
notebook
PyYAML
pandas
numpy
sklearn
python-dotenv
mlflow


### local packages -
-e . 
